// src/api/generated/types.ts
export type Nullish<T> = T | null | undefined
// ...
export type Models_HomeItem = {
    id: string
    type: string
    /**
     * options chosen by the user
     */
    options?: any
    /**
     * checks if it's still valid on the client
     */
    schemaVersion: number
}